Migrate Collection Management Service To TS
